+++ apps/ufjtbot/package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "lint": "eslint . --fix",
     "start": "node --experimental-json-modules build/main.js",
+    "watch": "pnpm build -- -w",
     "build": "swc src -d build -C jsc.parser.syntax=typescript -C module.type=es6 -C jsc.target=es2021 -C jsc.externalHelpers=true -C jsc.experimental.keepImportAssertions=true --no-swcrc",
     "docker:build": "cmd dockerized apps/ufjtbot/Dockerfile --exclude=__tests__/** .env.example | docker build --build-arg PACKAGE_PATH=apps/ufjtbot - -t mono-bot"
   },
