+++ apps/ufjtbot/src/github/pullRequest/closed.ts
@@ -1,7 +1,7 @@
 import { Context } from 'probot';
 import { logger } from '@ufjt-poc/logger';
 import { getErrorMessage, useSlackClient } from '../../utils.js';
-import { getChannelsFromRepository, getPullRequestMessages } from '../../requests.js';
+import { getChannelsFromRepository, getPullRequestMessages, removeConversation } from '../../requests.js';
 import mergedMessage from '../../slack/blocks/mergedMessage.js';
 
 export default async function closed({ payload }: Context<'pull_request.closed'>) {
@@ -36,9 +36,9 @@ export default async function closed({ payload }: Context<'pull_request.closed'>
         }),
         ts: ts
       });
+      await removeConversation({ channel, ts });
     }));
   } catch (err) {
     logger.error(getErrorMessage(err));
   }
-
 }
