+++ apps/ufjtbot/src/slack/blocks/mergedMessage.ts
@@ -7,7 +7,7 @@ export default function mergedMessage({ user, url, title, branch, state }: Merge
       'elements': [
         {
           'type': 'mrkdwn',
-          'text': `*Merged*: @${user} has ${state} ~<${url}|${title}>~ ${state === 'merged' ? `to ${branch}` : ''}`
+          'text': `@${user} has ${state} ~<${url}|${title}>~ ${state === 'merged' ? `to ${branch}` : ''}`
         }
       ]
     }
