+++ apps/ufjtbot/src/slack/commands/onUFJTCommand.ts
@@ -1,7 +1,6 @@
 import { AllMiddlewareArgs, SlackCommandMiddlewareArgs } from '@slack/bolt';
 import subscribeHandler from './subscribeHandler.js';
 import unsubscribeHandler from './unsubscribeHandler.js';
-import setReviewer from './setReviewer.js';
 
 export type OnUFJTCommand = SlackCommandMiddlewareArgs & AllMiddlewareArgs;
 
@@ -12,7 +11,7 @@ export type OnUFJTCommand = SlackCommandMiddlewareArgs & AllMiddlewareArgs;
  * @param respond
  */
 // eslint-disable-next-line complexity,max-lines-per-function,max-statements
-export default async function onUFJTCommand({ ack, command, respond, body, client }: OnUFJTCommand): Promise<void> {
+export default async function onUFJTCommand({ ack, command, respond }: OnUFJTCommand): Promise<void> {
   await ack();
   const [commandAction] = command.text.split(' ');
 
@@ -22,9 +21,6 @@ export default async function onUFJTCommand({ ack, command, respond, body, clien
       break;
     case 'unsub':
       await unsubscribeHandler({ command, respond });
-      break;
-    case 'mentions':
-      await setReviewer({ command, respond, body, client });
       break;
     default:
       break;
